Hide match amount on Direct grants rounds
